+++ crates/full-node/sov-stf-runner/tests/integration/helpers/mod.rs
@@ -1,2 +1,10 @@
+use sov_stf_runner::RawGenesisStateRoot;
+
 pub mod hash_stf;
 pub mod runner_init;
+
+const GENESIS_STATE_ROOT: [u8; 32] = [22; 32];
+
+pub fn genesis_state_root() -> RawGenesisStateRoot {
+    RawGenesisStateRoot(GENESIS_STATE_ROOT.to_vec())
+}